feat: add session id
Adds a session id parameter that is initialized once at startup and
sent with each request. This is an attempt to keep track of clients that
are anonymized for at least the length of an app run so ad frequency
settings are somewhat adhered to.

This also removes and cleans up several logging calls that were overly
verbose or incorrectly formatted.